backend/requirements.txt
Flask==2.2.3
Werkzeug==2.3.7
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.4.4
Flask-Cors==3.0.10
pymongo==4.3.3
python-dotenv==1.0.0
pydantic==2.0.2
email-validator==2.1.0.post1
langdetect==1.0.9
gunicorn==20.1.0
xgboost==1.7.3
scikit-learn==1.2.2
joblib==1.2.0
tldextract==3.4.0
dnspython==2.3.0
numpy==1.24.3
requests==2.28.2
google-generativeai>=0.6.0
google-cloud-translate==3.11.3
google-api-python-client==2.153.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
bcrypt==4.1.3
nltk==3.9.1
pandas==2.2.2
matplotlib==3.8.4 